fix: fixed the terminal output error

+++ inc/timer.hpp
@@ -6,6 +6,6 @@ using namespace std;
 
 struct s_emu_state;
 
-#define TIMER_CONFIG 0xFFFFFF10
+#define TIMER_CONFIG ((uint32_t)0xFFFFFF10)
 
 void timer_loop(s_emu_state* state);
